pyproject.toml
[project]
name = "hello-agent"
version = "0.1.0"
description = "Hello-world agent built on Norns."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "norns-sdk",
]

[project.scripts]
hello-worker = "hello.worker:main"
hello-client = "hello.client:main"

[tool.hatch.build.targets.wheel]
packages = ["hello"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
norns-sdk = { path = "../norns-sdk-python", editable = true }
